Validation changes

+++ src/main/java/com/sandeepbegudem/customer/payments/service/controller/CustomerController.java
@@ -2,7 +2,7 @@ package com.sandeepbegudem.customer.payments.service.controller;
 
 import com.sandeepbegudem.customer.payments.service.dto.CustomerPaymentsRequest;
 import com.sandeepbegudem.customer.payments.service.dto.CustomerResponse;
-import com.sandeepbegudem.customer.payments.service.entity.Customer;
+import com.sandeepbegudem.customer.payments.service.exception.CustomerNotFoundException;
 import com.sandeepbegudem.customer.payments.service.service.CustomerService;
 import com.sandeepbegudem.customer.payments.service.service.JwtService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -13,7 +13,6 @@ import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
-import java.util.Optional;
 
 @RestController
 @RequestMapping("/api/v1/customers")
@@ -32,42 +31,30 @@ public class CustomerController {
         this.authenticationManager = authenticationManager;
     }
 
-    @GetMapping
-    @PreAuthorize("hasAuthority('ADMIN')")
-    public ResponseEntity<List<CustomerResponse>> getAllCustomers(){
-
-        return new ResponseEntity<>(customerService.getAllCustomers(), HttpStatus.OK);
-    }
-
-//    @GetMapping
-//    @PreAuthorize("hasAuthority('ADMIN')")
-//    public ResponseEntity<List<CustomerPaymentsRequest>> retrieveAllCustomers(){
-//
-//        return new ResponseEntity<>(customerService.getAllCustomers(), HttpStatus.OK);
-//    }
-
     @PostMapping
     @PreAuthorize("hasAuthority('USER')")
     public ResponseEntity<CustomerResponse> insertCustomer(@RequestBody CustomerPaymentsRequest request){
-
         return new ResponseEntity<>(customerService.saveCustomer(request), HttpStatus.CREATED);
     }
 
     @GetMapping("/{id}")
     @PreAuthorize("hasAuthority('USER')")
-    public ResponseEntity<CustomerResponse> retrieveCustomerById(@PathVariable int id){
-
+    public ResponseEntity<CustomerResponse> retrieveCustomerById(@PathVariable Integer id) throws CustomerNotFoundException{
         return new ResponseEntity<>(customerService.customerById(id), HttpStatus.OK);
     }
 
     @ResponseStatus(HttpStatus.NO_CONTENT)
     @DeleteMapping("/{id}")
     @PreAuthorize("hasAuthority('ADMIN')")
-    public void deleteCustomerById(@PathVariable Integer id) {
-        if (id == null) {
-            throw new RuntimeException("id: " + " not found");
-        }
+    public void deleteCustomerById(@PathVariable Integer id) throws CustomerNotFoundException {
+        if (null != id) customerService.deleteCustomerById(id);
         else
-            customerService.deleteCustomerById(id);
+            throw new RuntimeException("id : " + id + " can't be null");
+    }
+
+    @GetMapping
+    @PreAuthorize("hasAuthority('ADMIN')")
+    public ResponseEntity<List<CustomerResponse>> getAllCustomers(){
+        return new ResponseEntity<>(customerService.getAllCustomers(), HttpStatus.OK);
     }
 }
